requirements.txt
aiofiles==24.1.0
fastapi==0.115.0
jinja2==3.1.4
markupsafe==2.1.3
matplotlib==3.9.2
numpy<2.0.0
pandas==2.2.2
pillow==11.0.0
python-multipart==0.0.19
starlette==0.38.6
torch==2.2.2
torchvision==0.17.2
transformers==4.46.3
uvicorn==0.30.6

